refactor(architecture): relocate transport constants out of server/
Resolves the #25 dependency-rule inversion: transport/ imported
ATTACHMENT_MAX_BYTES/MIME_SAFELIST from server/constants — a lower
layer reaching into the composition root.

- Move ATTACHMENT_MAX_BYTES, MIME_SAFELIST → infra/http/constants.ts
- Delete 11 duplicated/orphan constants from server/constants.ts
  (canonical copies already in infra/http, infra/tunnel, telegram)
- Repoint server/config.ts + config.test.ts to the infra locations
- server/constants.ts now holds only PILOT_VERSION (release-script
  hard reference) plus composition-root-specific symbols, with header
  comments pointing to each canonical home

Pure refactor — zero behavior change, constant values byte-identical.
tsc --noEmit clean, bun test 558/0. Every duplicate/orphan claim was
verified against its importers before deletion (no discrepancies).

Stacked on #26 (modifies sessions.ts, which #26 also changed). Refs #25.
